Bootstrap Confidence Intervals Linear Regression for Compositional Outcomes and Predictors

Implements the EM algorithm as described in Fiksel et al. (2020) for transformation-free linear regression for compositional outcomes and predictors.

codalm_ci(
  y,
  x,
  accelerate = TRUE,
  nboot = 500,
  ci_type = "perc",
  conf = 0.95,
  ...
)

Arguments

y
A matrix of compositional outcomes. Each row is an observation, and must sum to 1. If any rows do not sum to 1, they will be renormalized

x
A matrix of compositional predictors. Each row is an observation, and must sum to 1. If any rows do not sum to 1, they will be renormalized

accelerate
A logical variable, indicating whether or not to use the Squarem algorithm for acceleration of the EM algorithm. Default is TRUE

nboot
The number of bootstrap reptitions to use. Default is 500

ci_type
A character string with the type of bootstrap interval to be calculated. One of "norm", "perc", "basic", or "bca". See the documentation for boot.ci for more information. Default is "perc".

conf
A scalar between 0 and 1 containing the confidence level of the required intervals. Default is .95.

...
Additional arguments to pass to the boot function. If you want to use parallelize the bootstrapping, you can pass in the appropriate arguments here.

Value

A list, with ci_L and

Examples

require(ggtern)
data("WhiteCells", package = 'ggtern')
image <- subset(WhiteCells, Experiment == "ImageAnalysis")
image_mat <- as.matrix(image[,c("G", "L", "M")])
microscopic <- subset(WhiteCells, Experiment == "MicroscopicInspection")
microscopic_mat <- as.matrix(microscopic[,c("G", "L", "M")])
x  <- image_mat  / rowSums(image_mat)
y <- microscopic_mat / rowSums(microscopic_mat)
codalm_ci(y, x, nboot = 50, ci_type = "perc", conf = .95)

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> Warning: extreme order statistics used as endpoints

#> $ci_L
#>              [,1]         [,2]         [,3]
#> [1,] 9.426786e-01 1.154185e-02 9.550892e-04
#> [2,] 1.542272e-17 9.637253e-01 4.235204e-16
#> [3,] 9.604150e-18 8.208338e-08 8.946832e-01
#> 
#> $ci_U
#>            [,1]       [,2]         [,3]
#> [1,] 0.98599467 0.05516547 5.937150e-03
#> [2,] 0.03627470 1.00000000 2.574564e-08
#> [3,] 0.01291892 0.10531676 9.955661e-01
#>